# Add perf optimizations for path comparisons