This adds a few minor changes to improve the performance of path comparisons.

Prior to #28629, this change shaved about 60 seconds off of the owa controls build time, though following #28629 this only saves about 2 seconds.

```
Before:
Total time:  142.71s

After:
Total time:  140.65s
```