@rbuckton 2018-11-26
@weswigham: Can you take a look at this change today? Without the caching in #28629, this was a major performance bottleneck, so I can definitely see it coming back as an issue at some point in the future.